B0.1.4B - Center Diagonal Autonomous
Revamped the way PIDDriveRotate works. Now the angles that are input, must be based off of the starting configuration of the robot. Inputting 0 will make the robot face the same direction it began the autonomous routine (wherever it first zero'd itself). Completed the first versions of the adaptive center diagonal autonomous. Also optimized the velocity and acceleration values. Added the code to implement a usb camera and robot length adaptive between competition and practice robot.